restructure types to a separate folder
