+++ lib/util.js
@@ -1,4 +1,5 @@
 'use strict';
+
 const UPPER_LETTERS = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ';
 const LOWER_LETTERS = 'abcdefghijklmnopqrstyvwxyz';
 
@@ -23,6 +24,15 @@ const formatLength = (length) => {
   return length;
 };
 
+const checkLength = (value, path, length) => {
+  const len = value.length;
+  const { min, max } = length;
+  if (min && len < min) return [`Field "${path}" value is too short`];
+  if (max && len > max) {
+    return [`Field "${path}" exceeds the maximum length`];
+  }
+};
+
 const omit = (key, obj) => {
   const { [key]: omitted, ...rest } = obj;
   return [omitted, rest];
@@ -38,4 +48,5 @@ module.exports = {
   isFirstUpper,
   formatLength,
   isFirstLetter,
+  checkLength,
 };
